Altered code in 'PlayerControl.cpp'
Altered the code so that 'LevelBuilder' is passed through as an object so that collision can be run with each of the rendered cubes. However, the code has been commented out due to issues with the frame rate and the fact that the current method seems really inefficient. The frame rate drop also affects movement, so either find a better way to move or fix the frame rate to 60.