Fixing/Moving exisiting
Fixed the code for movement in 'PlayerControl.cpp' so it's not tied to the frame rate of the game.  Created a new .cpp called 'Cube.cpp'. This will be an instance of a cube that will be called by 'LevelBuilder.cpp' and will hold the 'CollisonManager' that is currently residing in 'PlayerControl.pp'.